requirements.txt
Flask
h5py
numpy
# tensorflow
tensorflow-gpu
sgfmill
